fixing bug with face rec
