+++ intelligencelayer/shared/models/common.py
@@ -6,30 +6,6 @@ import torch.nn as nn
 from utils.general import non_max_suppression
 import torch.nn.functional as F
 
-class Hardswish(nn.Module):
-    r"""Applies the hardswish function, element-wise, as described in the paper:
-    `Searching for MobileNetV3`_.
-    .. math::
-        \text{Hardswish}(x) = \begin{cases}
-            0 & \text{if~} x \le -3, \\
-            x & \text{if~} x \ge +3, \\
-            x \cdot (x + 3) /6 & \text{otherwise}
-        \end{cases}
-    Shape:
-        - Input: :math:`(N, *)` where `*` means, any number of additional
-          dimensions
-        - Output: :math:`(N, *)`, same shape as the input
-    Examples::
-        >>> m = nn.Hardswish()
-        >>> input = torch.randn(2)
-        >>> output = m(input)
-    .. _`Searching for MobileNetV3`:
-        https://arxiv.org/abs/1905.02244
-    """
-
-    def forward(self, input: torch.Tensor) -> torch.Tensor:
-        return F.hardswish(input)
-
 def autopad(k, p=None):  # kernel, padding
     # Pad to 'same'
     if p is None:
