+++ intelligencelayer/shared/process.py
@@ -23,7 +23,8 @@ from utils.general import (
     xyxy2xywh,
 )
 from utils.torch_utils import load_classifier, select_device, time_synchronized
-
+import torch.nn as nn
+from utils.activations import Hardswish
 
 class YOLODetector(object):
     def __init__(self, model_path: str, reso: int = 640, cuda: bool = False):
@@ -34,6 +35,13 @@ class YOLODetector(object):
         self.reso = (reso, reso)
         self.cuda = cuda
         self.model = attempt_load(model_path, map_location=self.device)
+
+        # Update model
+        for k, m in self.model.named_modules():
+            m._non_persistent_buffers_set = set()  # pytorch 1.6.0 compatability
+            if isinstance(m, self.models.common.Conv) and isinstance(m.act, nn.Hardswish):
+                m.act = Hardswish()  # assign activation
+
         self.names = (
             self.model.module.names
             if hasattr(self.model, "module")
